cleaning up the cleanup for accounts restructure;
